перенесены константы
