+++ js/util.js
@@ -32,54 +32,6 @@ const checkMaxLength = (string, maxLength) => {
 
 checkMaxLength('ПРОВЕРКА длины строки', 10);
 
-const NAMES = [
-  'Елена',
-  'Иван',
-  'Николай',
-  'Галина',
-  'Алексей',
-  'Нина',
-  'Виктория',
-  'Константин',
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const DESCRIPTION = [
-  'Пляж',
-  'Указатель',
-  'Море',
-  'На отдыхе',
-  'Веселые человечки',
-  'Хочу такую же',
-  'Два кусочека клубнички',
-  'Аперитив',
-  'Береговая охрана',
-  'Обувница',
-  'Дорога к морю',
-  'В поселке',
-  'Салат',
-  'Не ешь меня, я тебе еще пригожусь',
-  'Я отдыхаю',
-  'Путь домой',
-  'Хор местных жителей',
-  'Продается',
-  'Караул!!! Грабят!!!',
-  'Пальмы, пальмы, пальмы...',
-  'Обед',
-  'Закат',
-  'Краб',
-  'Флэшмоб',
-  'Переправа',
-];
-
 const shuffle = (count) => {
   const identifiers = Array.from({length: count}, (v, k) => k+1);
   for (let i = count - 1; i > 0; i--) {
@@ -93,4 +45,4 @@ const getRandomArrayElement = (elements) => {
   return elements[getRandomInt(0, elements.length - 1)];
 };
 
-export {getRandomInt, shuffle, getRandomArrayElement, DESCRIPTION, MESSAGES, NAMES};
+export {getRandomInt, shuffle, getRandomArrayElement};
